Create base project launcher
