+++ src/main/kotlin/com/jetbrains/rider/aspire/sessionHost/projectLaunchers/AspireProjectBaseLauncher.kt
@@ -1,0 +1,135 @@
+package com.jetbrains.rider.aspire.sessionHost.projectLaunchers
+
+import com.intellij.execution.executors.DefaultRunExecutor
+import com.intellij.execution.process.KillableProcessHandler
+import com.intellij.execution.process.ProcessAdapter
+import com.intellij.execution.process.ProcessEvent
+import com.intellij.execution.process.ProcessHandler
+import com.intellij.execution.process.ProcessOutputType
+import com.intellij.openapi.application.readAction
+import com.intellij.openapi.diagnostic.logger
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.Key
+import com.jetbrains.rd.util.lifetime.Lifetime
+import com.jetbrains.rider.aspire.generated.SessionModel
+import com.jetbrains.rider.aspire.sessionHost.SessionEvent
+import com.jetbrains.rider.aspire.sessionHost.SessionExecutableFactory
+import com.jetbrains.rider.aspire.sessionHost.SessionLogReceived
+import com.jetbrains.rider.aspire.sessionHost.SessionStarted
+import com.jetbrains.rider.aspire.sessionHost.SessionTerminated
+import com.jetbrains.rider.aspire.sessionHost.findBySessionProject
+import com.jetbrains.rider.aspire.sessionHost.hotReload.AspireProjectHotReloadConfigurationExtension
+import com.jetbrains.rider.aspire.util.decodeAnsiCommandsToString
+import com.jetbrains.rider.model.runnableProjectsModel
+import com.jetbrains.rider.projectView.solution
+import com.jetbrains.rider.run.configurations.RunnableProjectKinds
+import com.jetbrains.rider.run.configurations.RuntimeHotReloadRunConfigurationInfo
+import com.jetbrains.rider.run.configurations.launchSettings.LaunchSettingsJsonService
+import com.jetbrains.rider.run.pid
+import com.jetbrains.rider.runtime.DotNetExecutable
+import com.jetbrains.rider.runtime.DotNetRuntime
+import com.jetbrains.rider.runtime.RiderDotNetActiveRuntimeHost
+import com.jetbrains.rider.runtime.dotNetCore.DotNetCoreRuntime
+import kotlinx.coroutines.flow.MutableSharedFlow
+import java.nio.file.Path
+
+abstract class AspireProjectBaseLauncher : AspireProjectLauncherExtension {
+    companion object {
+        private val LOG = logger<AspireProjectBaseLauncher>()
+    }
+
+    protected abstract val hotReloadExtension: AspireProjectHotReloadConfigurationExtension
+
+    protected suspend fun getExecutable(sessionModel: SessionModel, project: Project): DotNetExecutable? {
+        val factory = SessionExecutableFactory.getInstance(project)
+        val executable = factory.createExecutable(sessionModel)
+        if (executable == null) {
+            LOG.warn("Unable to create executable for project: ${sessionModel.projectPath}")
+        }
+
+        return executable
+    }
+
+    protected fun getRuntime(executable: DotNetExecutable, project: Project): DotNetCoreRuntime? {
+        val runtime = DotNetRuntime.detectRuntimeForProject(
+            project,
+            RunnableProjectKinds.DotNetCore,
+            RiderDotNetActiveRuntimeHost.getInstance(project),
+            executable.runtimeType,
+            executable.exePath,
+            executable.projectTfm
+        )?.runtime as? DotNetCoreRuntime
+        if (runtime == null) {
+            LOG.warn("Unable to detect runtime for executable: ${executable.exePath}")
+        }
+
+        return runtime
+    }
+
+    protected fun subscribeToSessionEvents(
+        sessionId: String,
+        handler: ProcessHandler,
+        sessionEvents: MutableSharedFlow<SessionEvent>
+    ) {
+        handler.addProcessListener(object : ProcessAdapter() {
+            override fun startNotified(event: ProcessEvent) {
+                LOG.trace("Aspire session process started (id: $sessionId)")
+                val pid = when (event.processHandler) {
+                    is KillableProcessHandler -> event.processHandler.pid()
+                    else -> null
+                }
+                if (pid == null) {
+                    LOG.warn("Unable to determine process id for the session $sessionId")
+                    sessionEvents.tryEmit(SessionTerminated(sessionId, -1))
+                } else {
+                    sessionEvents.tryEmit(SessionStarted(sessionId, pid))
+                }
+            }
+
+            override fun onTextAvailable(event: ProcessEvent, outputType: Key<*>) {
+                val text = decodeAnsiCommandsToString(event.text, outputType)
+                val isStdErr = outputType == ProcessOutputType.STDERR
+                sessionEvents.tryEmit(SessionLogReceived(sessionId, isStdErr, text))
+            }
+
+            override fun processNotStarted() {
+                LOG.warn("Aspire session process is not started")
+                sessionEvents.tryEmit(SessionTerminated(sessionId, -1))
+            }
+        })
+    }
+
+    protected suspend fun enableHotReload(
+        executable: DotNetExecutable,
+        sessionProjectPath: Path,
+        launchProfile: String?,
+        lifetime: Lifetime,
+        project: Project
+    ): Pair<DotNetExecutable, ProcessAdapter?> {
+        val runnableProject = project.solution.runnableProjectsModel.findBySessionProject(sessionProjectPath)
+            ?: return executable to null
+
+        val hotReloadRunInfo = RuntimeHotReloadRunConfigurationInfo(
+            DefaultRunExecutor.EXECUTOR_ID,
+            project,
+            runnableProject,
+            executable.projectTfm,
+            null
+        )
+
+        val profile = if (!launchProfile.isNullOrEmpty()) {
+            val launchSettings = readAction {
+                LaunchSettingsJsonService.loadLaunchSettings(runnableProject)
+            }
+            launchSettings?.let { ls ->
+                ls.profiles?.get(launchProfile)
+            }
+        } else null
+
+        if (!hotReloadExtension.canExecute(lifetime, hotReloadRunInfo, profile)) {
+            return executable to null
+        }
+
+        return hotReloadExtension.execute(executable, lifetime, project)
+    }
+}
